requirements.txt
aniso8601==9.0.1
asn1crypto==1.5.1
attrs==23.1.0
base58==2.1.1
bitstring==3.1.9
blinker==1.6.2
cffi==1.15.1
click==8.1.3
coincurve==17.0.0
cryptography==36.0.2
Flask==2.3.2
flask-restx==1.1.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
json5==0.9.14
jsonschema==4.17.3
MarkupSafe==2.1.3
pycparser==2.21
pyln-bolt7==1.0.246
pyln-client==23.5
pyln-proto==23.5
pyrsistent==0.19.3
PySocks==1.7.1
pytz==2023.3
Werkzeug==2.3.6
